// src/models/ITRFailedPayment.model.ts
export type ITRFailedPaymentListResponse = {
    _id: string
    firstName: string
    middleName: string
    lastName: string
    adhaarNumber: string
    assesmentYear: string
    incomeSource: string
    fillingType: string
    mobileNumber: string
    emailId: string
    adhaarFrontPhotoUrl: string
    adhaarBackPhotoUrl: string
    panCardPhotoUrl: string
    banPassbookPhotoUrl: string
    otherDocuments: any[]
    distributorCode: string
    distributorId: string
    txnId: string
    txnObjectId: string
    uniqueTransactionId: string
    serialNumber: number
    srn: string
    totalPrice: number
    additionalCat2: number
    additionalCat3: number
    convinienceCharges: number
    basePrice: number
    baseCatAppied: string
    refundWalletAmountApplied: number
    rewardWalletAmountApplied: number
    paymentCategory: string[]
    appliedFrom: string
    version: string
    acknowledgementNumber: string
    acknowledgementPdf: string
    status: string
    txnStatus: string
    comments: string
    rejectionReason: string
    remark: string
    appliedByUserCurrentAddress: string
    appliedByType: string
    appliedById: string
    appliedOnDate: string
    appliedByName: string
    appliedByNumber: string
    appliedAsType: string
    assignedToName: string
    assignedToId: string
    assignedOnDate: string
    assignedByName: string
    assignedById: string
    verifiedByName: string
    verifiedById: string
    verifiedOnDate: string
    rejectedByName: string
    rejectedById: string
    rejectedOnDate: string
    generatedByName: string
    generatedById: string
    generatedOnDate: string
    completedByName: string
    completedById: string
    completedOnDate: string
    cancelledByName: string
    cancelledById: string
    cancelledOnDate: string
    moveToPendingByName: string
    moveToPendingById: string
    moveToPendingOnDate: string
    IsAgreedToTermsAndConditions: boolean
    retailerId: string
    retailerMobileNumber: string
    isDeleted: boolean
    createdAt: string
    updatedAt: string
  }